drm: kirin: Add suspend/resume hooks to send DPMS signals
Currently on suspend/resume, displays connected to the hikey
board won't ever go into sleep mode, as we never send the
DPMS signals to shutdown.

Thus this patch adds hooks for suspend/resume to do so,
following the style of the exynos driver.

NOTE: Suspending the display works fine, but for some reason
resuming the display is iffy. Can take a very long time to
actually happen (if it does at all). Still looking into it.

Feedback would be greatly appreciated!
Signed-off-by: John Stultz <john.stultz@linaro.org>